feat: moved the code in to headers
