+++ exercise-4/src/tcp_echo_client.cpp
@@ -6,30 +6,30 @@
 #include <sys/socket.h>
 #include <sys/types.h>
 #include <unistd.h>
+#include "socket_utilities.h"
+// void check_error(bool test, std::string error_message) {
+//   if (test) {
+//     std::cerr << error_message << "\n";
+//     exit(EXIT_FAILURE);
+//   }
+// }
 
-void check_error(bool test, std::string error_message) {
-  if (test) {
-    std::cerr << error_message << "\n";
-    exit(EXIT_FAILURE);
-  }
-}
+// int create_socket() {
+//   int sock = socket(AF_INET, SOCK_STREAM, 0);
+//   check_error(sock < 0, "Socket creation error\n");
+//   return sock;
+// }
 
-int create_socket() {
-  int sock = socket(AF_INET, SOCK_STREAM, 0);
-  check_error(sock < 0, "Socket creation error\n");
-  return sock;
-}
+// sockaddr_in create_address(const std::string &server_ip, int port) {
+//   sockaddr_in address;
+//   address.sin_family = AF_INET;
+//   address.sin_port = htons(port);
 
-sockaddr_in create_address(const std::string &server_ip, int port) {
-  sockaddr_in address;
-  address.sin_family = AF_INET;
-  address.sin_port = htons(port);
-
-  // Convert the server IP address to a binary format
-  auto err_code = inet_pton(AF_INET, server_ip.c_str(), &address.sin_addr);
-  check_error(err_code <= 0, "Invalid address/ Address not supported.\n");
-  return address;
-}
+//   // Convert the server IP address to a binary format
+//   auto err_code = inet_pton(AF_INET, server_ip.c_str(), &address.sin_addr);
+//   check_error(err_code <= 0, "Invalid address/ Address not supported.\n");
+//   return address;
+// }
 
 void connect_to_server(int sock, sockaddr_in &server_address) {
   auto err_code =
