+++ exercise-4/src/tcp_echo_server.cpp
@@ -3,20 +3,20 @@
 #include <sys/socket.h>
 #include <sys/types.h>
 #include <unistd.h>
+#include "socket_utilities.h"
+// template <typename T, typename S> void check_error(T test, S error_message) {
+//   if (test) {
+//     std::cerr << error_message << "\n";
+//     exit(EXIT_FAILURE);
+//   }
+// }
 
-template <typename T, typename S> void check_error(T test, S error_message) {
-  if (test) {
-    std::cerr << error_message << "\n";
-    exit(EXIT_FAILURE);
-  }
-}
-
-int create_socket() {
-  int my_sock;
-  my_sock = socket(AF_INET, SOCK_STREAM, 0);
-  check_error(my_sock < 0, "Socket creation error\n");
-  return my_sock;
-}
+// int create_socket() {
+//   int my_sock;
+//   my_sock = socket(AF_INET, SOCK_STREAM, 0);
+//   check_error(my_sock < 0, "Socket creation error\n");
+//   return my_sock;
+// }
 
 void set_socket_options(int sock, int opt) {
   auto err_code = setsockopt(sock, SOL_SOCKET, SO_REUSEADDR | SO_REUSEPORT,
@@ -24,13 +24,13 @@ void set_socket_options(int sock, int opt) {
   check_error(err_code < 0, "setsockopt() error\n");
 }
 
-sockaddr_in create_address(int port) {
-  sockaddr_in address;
-  address.sin_family = AF_INET;
-  address.sin_addr.s_addr = INADDR_ANY;
-  address.sin_port = htons(port);
-  return address;
-}
+// sockaddr_in create_address(int port) {
+//   sockaddr_in address;
+//   address.sin_family = AF_INET;
+//   address.sin_addr.s_addr = INADDR_ANY;
+//   address.sin_port = htons(port);
+//   return address;
+// }
 
 void bind_address_to_socket(int sock, sockaddr_in &address) {
   auto err_code = bind(sock, (sockaddr *)&address, sizeof(address));
